go/consensus/cometbft: Move query factories
