+++ go/consensus/cometbft/apps/governance/query_factory.go
@@ -1,0 +1,62 @@
+package governance
+
+import (
+	"context"
+	"fmt"
+
+	abciAPI "github.com/oasisprotocol/oasis-core/go/consensus/cometbft/api"
+	governanceState "github.com/oasisprotocol/oasis-core/go/consensus/cometbft/apps/governance/state"
+	"github.com/oasisprotocol/oasis-core/go/storage/mkvs"
+	"github.com/oasisprotocol/oasis-core/go/storage/mkvs/syncer"
+)
+
+// QueryFactory is the governance query factory.
+type QueryFactory struct {
+	state abciAPI.ApplicationQueryState
+}
+
+// NewQueryFactory returns a new governance query factory
+// backed by the given application state.
+func NewQueryFactory(state abciAPI.ApplicationQueryState) *QueryFactory {
+	return &QueryFactory{
+		state: state,
+	}
+}
+
+// QueryAt returns a governance query for a specific height.
+func (f *QueryFactory) QueryAt(ctx context.Context, height int64) (*Query, error) {
+	tree, err := abciAPI.NewImmutableStateAt(ctx, f.state, height)
+	if err != nil {
+		return nil, err
+	}
+	state := governanceState.NewImmutableState(tree)
+	query := NewQuery(state)
+	return query, nil
+}
+
+// LightQueryFactory is the governance light query factory.
+type LightQueryFactory struct {
+	rooter abciAPI.StateRooter
+	syncer syncer.ReadSyncer
+}
+
+// NewLightQueryFactory returns a new governance query factory
+// backed by a trusted state root provider and an untrusted read syncer.
+func NewLightQueryFactory(rooter abciAPI.StateRooter, syncer syncer.ReadSyncer) *LightQueryFactory {
+	return &LightQueryFactory{
+		rooter: rooter,
+		syncer: syncer,
+	}
+}
+
+// QueryAt returns a governance query for a specific height.
+func (f *LightQueryFactory) QueryAt(ctx context.Context, height int64) (*Query, error) {
+	root, err := f.rooter.StateRoot(ctx, height)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get state root: %w", err)
+	}
+	tree := mkvs.NewWithRoot(f.syncer, nil, root)
+	state := governanceState.NewImmutableState(tree)
+	query := NewQuery(state)
+	return query, nil
+}
